Extract bundle spawner code
